renamed to Ordered Covering Compressor
