+++ pacman/operations/algorithms_metadata.xml
@@ -483,8 +483,33 @@
             <token part="UnCompressedRoutingTablesGenerated">RoutingTablesGenerated</token>
         </outputs>
     </algorithm>
+    <algorithm name="OrderedCoveringCompressor">
+        <python_module>pacman.operations.router_compressors.ordered_covering_router_compressor.ordered_covering_compressor</python_module>
+        <python_class>OrderedCoveringCompressor</python_class>
+        <input_definitions>
+            <parameter>
+                <param_name>router_tables</param_name>
+                <param_type>MemoryRoutingTables</param_type>
+            </parameter>
+            <parameter>
+                <param_name>target_length</param_name>
+                <param_type>CompressionTargetSize</param_type>
+            </parameter>
+        </input_definitions>
+        <required_inputs>
+            <param_name>router_tables</param_name>
+        </required_inputs>
+        <optional_inputs>
+            <token>RoutingTablesPreCompressed</token>
+            <param_name>target_length</param_name>
+        </optional_inputs>
+        <outputs>
+            <param_type>MemoryCompressedRoutingTables</param_type>
+        </outputs>
+    </algorithm>
     <algorithm name="MundyRouterCompressor">
-        <python_module>pacman.operations.router_compressors.mundys_router_compressor.routing_table_condenser</python_module>
+        <!-- DEPRECATED Use OrderedCoveringRouterCompressor !-->
+        <python_module>pacman.operations.router_compressors.ordered_covering_router_compressor.ordered_covering_compressor</python_module>
         <python_class>MundyRouterCompressor</python_class>
         <input_definitions>
             <parameter>
