+++ pacman/operations/router_compressors/ordered_covering_router_compressor/remove_default_routes.py
