+++ unittests/operations_tests/router_compressor_tests/test_compressors.py
@@ -23,8 +23,8 @@ from pacman.operations.router_compressors.pair_compressor import (
     PairCompressor)
 from pacman.operations.router_compressors.unordered_compressor import (
     UnorderedCompressor)
-from pacman.operations.router_compressors.mundys_router_compressor.\
-    routing_table_condenser import (MundyRouterCompressor)
+from pacman.operations.router_compressors.ordered_covering_router_compressor \
+    import OrderedCoveringCompressor
 
 
 class MyTestCase(unittest.TestCase):
@@ -67,8 +67,8 @@ class MyTestCase(unittest.TestCase):
         compressed_tables = compressor(self.original_tables)
         self.check_compression(compressed_tables)
 
-    def test_mundy_compressor(self):
-        compressor = MundyRouterCompressor()
+    def test_ordered_covering_compressor(self):
+        compressor = OrderedCoveringCompressor()
         compressed_tables = compressor(self.original_tables)
         self.check_compression(compressed_tables)
